Add MemoryChecker.php
MemoryChecker lets you specify a soft memory limit as a ratio of the
memory_limit specified in your php.ini.

MemoryChecker is a singleton that will only throw an exception once per
process, so you can call it as many times as you want knowing that it
won't throw redundant exceptions.